test suite expanded to included Menu and Item classes
