+++ lib/takeaway.rb
@@ -1,3 +1,5 @@
+require_relative 'menu'
+
 class Order
 
   attr_reader :menu
@@ -73,24 +75,6 @@ class Order
   def order_total
     order_sum = @menu.map { |h| (h[:quant].to_f * h[:price].to_f).round(2) }
     "TOTAL: £#{order_sum.sum}"
-  end
-
-end
-
-class Menu
-  FULL_MENU = [
-    { :item => "Egg fried rice", :price => 4.99 },
-    { :item => "Steamed rice", :price => 4.49 },
-    { :item => "Sesame prawn toast", :price => 7.99 },
-    { :item => "Salt & pepper pork ribs", :price => 9.95 },
-    { :item => "Cantonese lemon chicken", :price => 8.95 },
-    { :item => "Lamb in black pepper sauce", :price => 7.99 }
-  ].freeze
-
-  attr_reader :import
-
-  def initialize
-    @import = FULL_MENU.each { |h| h[:quant] = 0 }
   end
 
 end
