+++ spec/takeaway_spec.rb
@@ -1,4 +1,4 @@
-require "takeaway"
+require 'takeaway'
 
 describe Order do
